small fixes

+++ app/api/decks/route.ts
@@ -16,10 +16,10 @@ export const POST = async (req: Request) => {
 }
 
 export const PUT = async (req: Request) => {
-     const {id, deckname, cards } = await req.json()
+     const {id, deckname, cards, owner } = await req.json()
      try {
         console.log("updating deck")
-        const { error } =  await supabase.from("decktable").update({deckname: deckname, cards: cards}).eq("id", id)
+        const { error } =  await supabase.from("decktable").update({deckname: deckname, cards: cards}).eq("id", id).eq("owner",owner)
         return new Response(JSON.stringify(error), {status: 201})
      } catch (error) {
         console.log(error)
@@ -27,11 +27,11 @@ export const PUT = async (req: Request) => {
 }
 
 export const DELETE = async (req: Request) => {
-    const { id } = await req.json()
+    const { id, owner } = await req.json()
     console.log(id)
     try {
         console.log("deleting deck")
-        const { error } = await supabase.from("decktable").delete().eq("id", id)
+        const { error } = await supabase.from("decktable").delete().eq("id", id).eq("owner", owner)
         return new Response(JSON.stringify(error), {status: 201})
     } catch (error) {
         console.log(error)
